Flutter/whatz_up_functions/functions/requirements.txt
firebase_functions~=0.1.0
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
firebase-admin==6.2.0